+++ src/core/contracts/ModuloEditorBuilder.ts
@@ -1,5 +1,5 @@
 import type {EditorPreset} from "../../presets";
-import type {EditorDomResolver} from "../../dom/contracts";
+import type {EditorDomInitializer, EditorDomResolver} from "../../dom/contracts";
 import type {EditorInputAdapter} from "../../input";
 import type {EditorOutputAdapter} from "../../output";
 import type {TextareaBridge} from "../../textarea";
@@ -11,6 +11,8 @@ import type {EditorCommand} from "../../commands";
 
 export interface ModuloEditorBuilder {
     usePreset(preset: EditorPreset): this;
+    fromTextarea(input: string | HTMLTextAreaElement): this;
+    withDomInitializer(domInitializer: EditorDomInitializer): this;
     withDomResolver(domResolver: EditorDomResolver): this;
     withInput(input: EditorInputAdapter): this;
     withOutput(output: EditorOutputAdapter): this;