+++ src/dom/index.ts
@@ -1,1 +1,2 @@
 export { DefaultEditorDomResolver } from "./DefaultEditorDomResolver";
+export {DefaultEditorDomInitializer} from "./DefaultEditorDomInitializer";